Add checks to return a failure if more or less than 2xnumbers are received
In some cases the tools are receiving more than 2xnumbers which causing
the workflow to spin.
Example:
[AGENT]
Calling tools: calculator_multiply
Tool's input: {"text": "60 * 39"}
Tool's response:
The product of 60 * 39 is 2340
------------------------------
2025-05-13 09:39:52,286 - aiq.agent.react_agent.agent - INFO -
------------------------------
[AGENT]
Agent input: What is 9 times 7, and is it greater than the current number of seconds since midnight?
Agent's thoughts:
Thought: Now that I have the product of 60 and 39, I can continue with the calculation to get the total number of seconds since midnight. I will add 2340 to 32400, then add 32.

Action: calculator_subtract
Action Input: {'text': '32400 + 2340 + 32'}
------------------------------
2025-05-13 09:39:52,291 - aiq.agent.react_agent.agent - INFO -
------------------------------
[AGENT]
Calling tools: calculator_subtract
Tool's input: {"text": "32400 + 2340 + 32"}
Tool's response:
The result of 32400 - 2340 is 30060
------------------------------
2025-05-13 09:39:56,498 - aiq.agent.react_agent.agent - INFO -
------------------------------
[AGENT]
Agent input: What is 9 times 7, and is it greater than the current number of seconds since midnight?
Agent's thoughts:
Thought: It seems there was an error in the previous calculation. The correct calculation should be 32400 + 2340 + 32. I will redo the calculation.

Signed-off-by: Anuradha Karuppiah <anuradhak@nvidia.com>