Enhance error handling in MCP tool invocation (#263)
Closes https://github.com/NVIDIA/AIQToolkit/issues/262

This PR handles cases when an MCP tool is invoked and fails because of invalid inputs. Previously, invalid inputs trigger exceptions and the entire agent exits, returning an error message. The resolution in this PR introduces a configuration setting in the MCP tool that propagates the exception message back to the agent, giving it an opportunity to recover. 

This exception can be observed with the `simple_mcp` example. Without this improvement, the `react_agent` fails to produce valid outputs after a single parsing failure. Note, this example will need to be run several times due to non-deterministic behavior. The `react_agent` intermittently produces invalid inputs for the MCP tool.

## By Submitting this PR I confirm:
- I am familiar with the [Contributing Guidelines](https://github.com/NVIDIA/AIQToolkit/blob/develop/docs/source/resources/contributing.md).
- We require that all contributors "sign-off" on their commits. This certifies that the contribution is your original work, or you have rights to submit it under the same license, or a compatible license.
  - Any contribution which contains commits that are not Signed-Off will not be accepted.
- When the PR is ready for review, new or existing tests cover these changes.
- When the PR is ready for review, the documentation is up to date with these changes.

Authors:
  - Matthew Penn (https://github.com/mpenn)

Approvers:
  - Anuradha Karuppiah (https://github.com/AnuradhaKaruppiah)

URL: https://github.com/NVIDIA/AIQToolkit/pull/263