Update `OpenAIModelConfig` to support `stream_usage` option (#328)
When using any of our agents and the OpenAI provider, token metrics are not reported to telemetry providers because the option `stream_usage` defaults to False in the OpenChatModel object.

This sets the default to be `True`, but leaves the option for users to override it as false with extra parameters.

Overhead should be minimal.

## By Submitting this PR I confirm:
- I am familiar with the [Contributing Guidelines](https://github.com/NVIDIA/AIQToolkit/blob/develop/docs/source/resources/contributing.md).
- We require that all contributors "sign-off" on their commits. This certifies that the contribution is your original work, or you have rights to submit it under the same license, or a compatible license.
  - Any contribution which contains commits that are not Signed-Off will not be accepted.
- When the PR is ready for review, new or existing tests cover these changes.
- When the PR is ready for review, the documentation is up to date with these changes.

Authors:
  - Michael Demoret (https://github.com/mdemoret-nv)

Approvers:
  - Yuchen Zhang (https://github.com/yczhang-nv)

URL: https://github.com/NVIDIA/AIQToolkit/pull/328