Account for the "required" list in the mcp_input_schema
If no default is defined in the schema, and the field isn’t in
the "required" list, it should default to None (i.e., optional).

Earlier logic marked the field as required (using ...) unless
"default" is explicitly set i.e. it would mark optional fields
with no default as required.

Signed-off-by: Anuradha Karuppiah <anuradhak@nvidia.com>