Ensure Profiler Agent Example Uses Provided Default Project Config (#465)
This pull request updates the profiler agent to run out of the box by requiring the user to provide a default_project_name, rather than relying on the hardcoded "default" value, which could lead to runtime errors. The example now sets 'simple_calculator' as the default project name, as it aligns with the simple_calculator_observability example used in the tests. The implementation uses the project name specified in the user query, and falls back to the configured default_project_name when none is provided.
Closes [AIQ-1577]( https://jirasw.nvidia.com/browse/AIQ-1577)

## By Submitting this PR I confirm:
- I am familiar with the [Contributing Guidelines](https://github.com/NVIDIA/NeMo-Agent-Toolkit/blob/develop/docs/source/resources/contributing.md).
- We require that all contributors "sign-off" on their commits. This certifies that the contribution is your original work, or you have rights to submit it under the same license, or a compatible license.
  - Any contribution which contains commits that are not Signed-Off will not be accepted.
- When the PR is ready for review, new or existing tests cover these changes.
- When the PR is ready for review, the documentation is up to date with these changes.

Authors:
  - Eric Evans II (https://github.com/ericevans-nv)

Approvers:
  - Yuchen Zhang (https://github.com/yczhang-nv)

URL: https://github.com/NVIDIA/NeMo-Agent-Toolkit/pull/465