Pin `datasets` version in toplevel `pyproject.toml`
`uv`'s solver incorrectly chooses wildly different `datasets` versions based on platform.
On linux, 4.0.0 is selected while on non-linux systems an older version (2.14.X) was chosen.

With the recent release of pyarrow 21, they removed deprecated code that the older
version of datasets was dependent upon.

This pins datasets accordingly so the solver doesn't yield package solves that vary so much.

Signed-off-by: Will Killian <wkillian@nvidia.com>