Timezone Support for `datetime` Tool and Normalize Response Time Handling (#660)
## Description
<!-- Note: The pull request title will be included in the CHANGELOG. -->
<!-- Provide a standalone description of changes in this PR. -->

- Added timezone support for the `datetime_tools`, so now it will
automatically extract the `x-timezone` from the headers from context,
which will be used to format the current time.
- If the `x-timezone` is not present in the headers, the fallback behavior is checked
    - The documentations should also be clear now.

- Fixed a small issue in
`NeMo-Agent-Toolkit/src/nat/builder/user_interaction_manager.py`. Before
the timestamp is in UTC (as seen by the `Z` suffix in the timestamp),
but the timestamp is in the local timezone, which is most likely not in
UTC timezone.

- Made a tiny change for the `NeMo-Agent-Toolkit-UI` repository. So now
the front-end also stores the timezone in its header to
be sent to the backend.
- The change:
[https://github.com/NVIDIA/NeMo-Agent-Toolkit-UI/pull/38](https://github.com/NVIDIA/NeMo-Agent-Toolkit-UI/pull/38)

Closes #445

## By Submitting this PR I confirm:
- I am familiar with the [Contributing
Guidelines](https://github.com/NVIDIA/NeMo-Agent-Toolkit/blob/develop/docs/source/resources/contributing.md).
- We require that all contributors "sign-off" on their commits. This
certifies that the contribution is your original work, or you have
rights to submit it under the same license, or a compatible license.
- Any contribution which contains commits that are not Signed-Off will
not be accepted.
- When the PR is ready for review, new or existing tests cover these
changes.
- When the PR is ready for review, the documentation is up to date with
these changes.

---------

Signed-off-by: Daniel Wang <daniewang@nvidia.com>