docs: use https GitHub URL for easier installation and contribution
The SSH protocol for GitHub URLs requires setting up an SSH key, even for public repositories, which increases the barrier to entry for quick installation. In contrast, the HTTPS version of GitHub URLs does not require authentication for public repositories, making it easier to get started. It also improves the contribution workflow for developers who prefer using Git over HTTPS.

Signed-off-by: Mengdi Gao <mengdig@nvidia.com>